Use simpler wireit config for shared scripts
