+++ packages/internal-scripts/src/index.js
@@ -1,2 +1,2 @@
-export { build } from './build.js';
-export { buildCDN } from './build-cdn.js';
+export { build } from './tasks/build.js';
+export { buildCDN } from './tasks/build-cdn.js';
